add(Object item) This method adds a new item to the end of the linked list.
First, a new node is created with the specified element and links to the next and previous nodes. Then it checks if the list is empty, and if so, the new node becomes the head. If the list is not empty, then the link to the next node of the last element is updated to the new node. Then the new node becomes the last element (tail), and the variable size increases by one.

Finally, the method returns true, which indicates the successful addition of the element.