add(Object element, int index) This method adds a new item to the specified position in the linked list.
First, it is checked whether the index is within acceptable limits. If the index is outside the size of the list, an IndexOutOfBoundsException is thrown.

Then it checks whether the item is added to the end of the list, if so, the method calls the add(Object item) method, which adds the item to the end of the list.

If the element is not added to the end of the list, then the method passes through the list to the node that is at the index position. Then a new node is created with the specified element, and its references to the next and previous nodes are set to the current node and the previous node, respectively. Then the references to the next and previous nodes of the current node are updated to the new node, and the size variable is incremented by one.