get(int index) is a method that returns the list item located at the specified index position.
The method first checks if the index value is within the valid index range for the current list, and if not, throws an IndexOutOfBoundsException.

The method then determines whether the item is closer to the beginning or the end of the list to start searching for the item from the end that is closer to the target position. If the index is in the first half of the list, the method uses a for loop to move through the list from the beginning to the target index. If the index is in the second half of the list, however, the method starts at the end of the list and then moves in the opposite direction to the target position.

Finally, the method returns the value of the date field of the node at the specified position.

Translated with www.DeepL.com/Translator (free version)